feat: remove code for unsupported version v6 of Guzzle
