+++ .changes/nextrelease/remove-guzzlev6-support.json
@@ -1,0 +1,7 @@
+[
+  {
+    "type": "enhancement",
+    "category": "Handler",
+    "description": "Remove code for unsupported version v6 of Guzzle and moved Aws\\Handler\\GuzzleV6\\GuzzleHandler to an unversioned Aws\\Handler\\Guzzle\\GuzzleHandler."
+  }
+]
